requirements.txt
requests>=2.28.0
beautifulsoup4>=4.12.2
faiss-cpu>=1.7.4
numpy>=1.24.0
transformers>=4.29.2
accelerate>=0.20.3
huggingface-hub>=0.15.1
python-dotenv>=1.0.0
tqdm>=4.65.0
streamlit
